authorization-service-server/src/test/java/server/infrastructure/azure/ad/AzureADUserPoolTest.java
package server.infrastructure.azure.ad;

import authorization.server.core.UserPoolException;
import authorization.server.infrastructure.azure.ad.AzureADUserPool;
import com.microsoft.aad.msal4j.*;
import lombok.extern.slf4j.Slf4j;
import org.mockito.*;
import org.testng.Assert;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.DataProvider;
import org.testng.annotations.Test;

import javax.naming.AuthenticationException;
import java.util.Collections;
import java.util.Set;
import java.util.concurrent.CompletableFuture;
import java.util.concurrent.ExecutionException;

@Slf4j
public class AzureADUserPoolTest {

    @Mock
    private PublicClientApplication mockPublicClientApplication;
    @Mock
    private ConfidentialClientApplication mockConfidentialClientApplication;
    @Mock
    private IAuthenticationResult mockAuthenticationResult;
    @Captor
    private ArgumentCaptor<UserNamePasswordParameters> userPasswordParametersCaptor;
    @Captor
    private ArgumentCaptor<ClientCredentialParameters> clientCredentialParametersCaptor;

    private AzureADUserPool azureAdUserPool;

    private final String clientId = System.getenv("AZURE_CLIENT_ID");
    private final String clientSecret = System.getenv("AZURE_CLIENT_SECRET");
    private final String tenantId = System.getenv("AZURE_TENANT_ID");
    private final String authority = "https://login.microsoftonline.com/"+tenantId+"/";

    @BeforeMethod
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        log.info("clientId:"+clientId);
        log.info("client Secret:"+clientSecret);
        log.info("Tenant:"+tenantId);
        azureAdUserPool = new AzureADUserPool(clientId, clientSecret, authority);
    }

    @Test
    public void testAuthenticate() throws Exception {
        String expectedToken = "id_token";

        // Set up the mock behavior for acquiring a token
        CompletableFuture<IAuthenticationResult> future = CompletableFuture.completedFuture(mockAuthenticationResult);
        Mockito.when(mockPublicClientApplication.acquireToken(Mockito.any(UserNamePasswordParameters.class))).thenReturn(future);
        Mockito.when(mockAuthenticationResult.idToken()).thenReturn(expectedToken);
        AzureADUserPool mockazureAdUserPool = new AzureADUserPool(clientId, clientSecret, authority, mockPublicClientApplication);

        String actualToken = mockazureAdUserPool.authenticate("foo", "bar");

        Assert.assertNotNull(actualToken);
        Assert.assertEquals(actualToken, expectedToken);
        Mockito.verify(mockPublicClientApplication).acquireToken(userPasswordParametersCaptor.capture());
        UserNamePasswordParameters capturedParams = userPasswordParametersCaptor.getValue();
        Assert.assertEquals(capturedParams.username(), "foo");
    }


    @Test(expectedExceptions = UserPoolException.class, expectedExceptionsMessageRegExp = "Authentication failed for user .*")
    public void testAuthenticateFailure() throws Exception {
        // Simulate an authentication failure
        CompletableFuture<IAuthenticationResult> failedFuture = new CompletableFuture<>();
        failedFuture.completeExceptionally(new AuthenticationException("Failed to authenticate"));
        Mockito.when(mockPublicClientApplication.acquireToken(Mockito.any(UserNamePasswordParameters.class))).thenReturn(failedFuture);

        azureAdUserPool.authenticate("foo", "bar");
    }

    @Test
    public void testAcquireTokenWithClientCredentials() throws Exception {
        Set<String> scopes = Collections.singleton("https://graph.microsoft.com/.default");
        String expectedPrefix = "eyJ0eXAiOiJKV1QiLCJub25jZSI6I";

        // Set up the mock behavior for client credentials flow
        CompletableFuture<IAuthenticationResult> future = CompletableFuture.completedFuture(mockAuthenticationResult);
        Mockito.when(mockConfidentialClientApplication.acquireToken(Mockito.any(ClientCredentialParameters.class))).thenReturn(future);
        Mockito.when(mockAuthenticationResult.accessToken()).thenReturn(expectedPrefix);

        String actualToken = azureAdUserPool.acquireTokenWithClientCredentials(scopes);

        Assert.assertNotNull(actualToken);
        Assert.assertTrue(actualToken.startsWith(expectedPrefix));
/* need to analyse and fix this later
        Mockito.verify(mockConfidentialClientApplication).acquireToken(clientCredentialParametersCaptor.capture());
        ClientCredentialParameters capturedParams = clientCredentialParametersCaptor.getValue();
        Assert.assertEquals(capturedParams.scopes(), scopes); */
    }

    @Test(expectedExceptions = UnsupportedOperationException.class, expectedExceptionsMessageRegExp = "Azure AD requires self-service password reset or admin intervention to change password.")
    public void testChangePasswordUnsupported() throws UserPoolException {
        // Azure AD does not support changing passwords directly through MSAL
        azureAdUserPool.changePassword("foo", "oldPass", "newPass");
    }

    @DataProvider
    public Object[][] authenticationExceptionSamples() {
        return new Object[][]{
                {ExecutionException.class}, {InterruptedException.class}
        };
    }

    @Test(dataProvider = "authenticationExceptionSamples", expectedExceptions = UserPoolException.class)
    public void testAuthenticateWithExceptions(Class<? extends Exception> exceptionClass) throws Exception {
        CompletableFuture<IAuthenticationResult> failedFuture = new CompletableFuture<>();
        Exception simulatedException;

        try {
            // Try to create an instance with a String constructor
            simulatedException = exceptionClass.getConstructor(String.class).newInstance("Simulated exception");
        } catch (NoSuchMethodException e) {
            // If no String constructor, use a Throwable constructor instead
            simulatedException = exceptionClass.getConstructor(Throwable.class).newInstance(new Throwable("Simulated exception"));
        }

        failedFuture.completeExceptionally(simulatedException);
        Mockito.when(mockPublicClientApplication.acquireToken(Mockito.any(UserNamePasswordParameters.class))).thenReturn(failedFuture);

        azureAdUserPool.authenticate("foo", "bar");
    }


}
